lib:sw_apps: Support memory tests for memory size more than 4 GB
support to run test for memory size more than 4GB by changing
data type from u32 to u64.


Acked-for-series:  Mubin Sayyed <mubin.sayyed@amd.com>

+++ lib/sw_apps/memory_tests/src/memory_config.h
@@ -12,7 +12,7 @@ struct memory_range_s {
     char8 *name;
     char8 *ip;
     UINTPTR base;
-    u32 size;
+    u64 size;
 };
 
 #define UPPER_4BYTES_MASK	0xFFFFFFFF00000000LL
